third_party/WebKit/LayoutTests/media/video-controls-zoomed.html

<!DOCTYPE html>
<title>Test controls with zooming.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script src="media-controls.js"></script>
<style>
video {
    margin: 50px;
}
</style>
<video controls></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_true(video.controls);

    video.oncanplaythrough = t.step_func(function() {
        setTimeout(function() {
            document.documentElement.style.zoom = "150%";
            video.onplaying = t.step_func_done(function() {
                assert_false(video.paused);
            });

            // Find the play button and click the middle of its bounding box.
            var playCoords = mediaControlsButtonCoordinates(video, "play-button");
            // Apply the page zoom value to the coordinates because
            // getBoundingClientRect() used in mediaControlsButtonCoordinates() doesn't do it.
            var clickX = playCoords[0] * 1.5;
            var clickY = playCoords[1] * 1.5;

            eventSender.mouseMoveTo(clickX, clickY);
            eventSender.mouseDown();
            eventSender.mouseUp();
        }, 50);
    });

    video.src = findMediaFile("video", "content/test");
});
</script>
